update auth service to user service communication, single get user function, using dto in service layer
