+++ routes/user.route.js
@@ -1,15 +1,15 @@
 const userController = require("../controllers/user.controller");
-const checkLogIn = require("../middlewares/auth.middleware");
+const checkLogin = require("../middlewares/auth.middleware");
 const express = require("express");
 const router = express.Router();
 
-router.route("/").get(checkLogIn,userController.getAllUser);
+router.route("/").get(checkLogin,userController.getAllUser);
 
 router
   .route("/:username")
-  .get(checkLogIn,userController.getUserByUsername)
-  .put(checkLogIn,userController.updateUserPasswordByUsername)
-  .delete(checkLogIn,userController.deleteUserByUsername);
+  .get(checkLogin,userController.getUserByUsername)
+  .put(checkLogin,userController.updateUserPasswordByUsername)
+  .delete(checkLogin,userController.deleteUserByUsername);
 
 
 module.exports = router;
